# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Recruiting OS API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type Seniority = (typeof Seniority)[keyof typeof Seniority];

export const Seniority = {
  Intern: "Intern",
  Junior: "Junior",
  Mid: "Mid",
  Senior: "Senior",
  Lead: "Lead",
  Principal: "Principal",
  Director: "Director",
  VP: "VP",
} as const;

export type ApplicationStage =
  (typeof ApplicationStage)[keyof typeof ApplicationStage];

export const ApplicationStage = {
  Sourced: "Sourced",
  Contacted: "Contacted",
  Screened: "Screened",
  Interview: "Interview",
  Offer: "Offer",
  Hired: "Hired",
  Rejected: "Rejected",
} as const;

export interface Job {
  id: number;
  title: string;
  description: string;
  location: string;
  seniority: Seniority;
  mustHaveSkills: string[];
  createdAt: string;
  updatedAt: string;
}

export interface CreateJobBody {
  title: string;
  description: string;
  location: string;
  seniority: Seniority;
  mustHaveSkills: string[];
}

export interface Candidate {
  id: number;
  name: string;
  email: string;
  linkedIn?: string | null;
  summary?: string | null;
  skills: string[];
  createdAt: string;
  updatedAt: string;
}

export interface CreateCandidateBody {
  name: string;
  email: string;
  linkedIn?: string | null;
  summary?: string | null;
  skills: string[];
}

export interface Application {
  id: number;
  jobId: number;
  candidateId: number;
  stage: ApplicationStage;
  notes?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface ApplicationWithDetails {
  id: number;
  jobId: number;
  candidateId: number;
  stage: ApplicationStage;
  notes?: string | null;
  createdAt: string;
  updatedAt: string;
  job: Job;
  candidate: Candidate;
}

export interface CreateApplicationBody {
  jobId: number;
  candidateId: number;
  stage: ApplicationStage;
  notes?: string | null;
}

export interface UpdateApplicationBody {
  stage: ApplicationStage;
  notes?: string | null;
}

export type ProviderType = (typeof ProviderType)[keyof typeof ProviderType];

export const ProviderType = {
  native_openai: "native_openai",
  custom_webhook: "custom_webhook",
  twin_webhook: "twin_webhook",
} as const;

export type WorkflowStepName =
  (typeof WorkflowStepName)[keyof typeof WorkflowStepName];

export const WorkflowStepName = {
  job_understanding: "job_understanding",
  candidate_matching: "candidate_matching",
  shortlist_generation: "shortlist_generation",
  sourcing_later: "sourcing_later",
} as const;

export interface AgentProviderRecord {
  id: number;
  name: string;
  type: ProviderType;
  baseUrl?: string | null;
  webhookUrl?: string | null;
  apiKeyEncryptedPlaceholder?: string | null;
  enabled: boolean;
  createdAt: string;
  updatedAt: string;
}

export interface CreateProviderBody {
  name: string;
  type: ProviderType;
  baseUrl?: string | null;
  webhookUrl?: string | null;
  apiKeyPlaceholder?: string | null;
  enabled?: boolean;
}

export interface ToggleProviderBody {
  enabled: boolean;
}

export interface ConnectionTestResult {
  ok: boolean;
  error?: string | null;
  latencyMs?: number | null;
}

export interface ProviderStepSetting {
  id: number;
  workflowStep: WorkflowStepName;
  providerId: number;
  enabled: boolean;
}

export interface ProviderStepSettingWithProvider {
  id: number;
  workflowStep: WorkflowStepName;
  providerId: number;
  enabled: boolean;
  provider: AgentProviderRecord;
}

export interface UpsertStepSettingBody {
  workflowStep: WorkflowStepName;
  providerId: number;
  enabled: boolean;
}

export interface RunWorkflowBody {
  jobId: number;
}

export type RunStatus = (typeof RunStatus)[keyof typeof RunStatus];

export const RunStatus = {
  pending: "pending",
  running: "running",
  completed: "completed",
  failed: "failed",
} as const;

export type StepStatus = (typeof StepStatus)[keyof typeof StepStatus];

export const StepStatus = {
  pending: "pending",
  running: "running",
  completed: "completed",
  failed: "failed",
} as const;

export interface AgentRun {
  id: number;
  jobId: number;
  status: RunStatus;
  createdAt: string;
  updatedAt: string;
}

export interface AgentLog {
  id: number;
  runId: number;
  step: string;
  input?: unknown | null;
  output?: unknown | null;
  status: StepStatus;
  createdAt: string;
}

export interface AgentRunWithLogs {
  id: number;
  jobId: number;
  status: RunStatus;
  createdAt: string;
  updatedAt: string;
  logs: AgentLog[];
}

export type Recommendation =
  (typeof Recommendation)[keyof typeof Recommendation];

export const Recommendation = {
  Strong_Yes: "Strong Yes",
  Yes: "Yes",
  Maybe: "Maybe",
  No: "No",
} as const;

export interface CandidateEvaluation {
  id: number;
  runId: number;
  jobId: number;
  candidateId: number;
  score: number;
  strengths: string[];
  gaps: string[];
  risks: string[];
  recommendation: Recommendation;
  candidate: Candidate;
  createdAt: string;
}

export interface JobInsight {
  id: number;
  runId: number;
  jobId: number;
  mustHaveSkills: string[];
  seniority: string;
  evaluationCriteria: string[];
  idealCandidateProfile: string;
  createdAt: string;
}

export interface ShortlistEntry {
  candidateId: number;
  candidateName: string;
  whyRelevant: string;
  keyRisks: string;
  finalRecommendation: string;
}

export interface Shortlist {
  id: number;
  runId: number;
  jobId: number;
  rankedCandidateIds: number[];
  summaries: ShortlistEntry[];
  createdAt: string;
}

export interface JobWorkflowResult {
  run: AgentRun;
  insight?: JobInsight | null;
  evaluations: CandidateEvaluation[];
  shortlist?: Shortlist | null;
  logs: AgentLog[];
}

export interface StageCount {
  stage: ApplicationStage;
  count: number;
}

export interface PipelineSummary {
  totalJobs: number;
  totalCandidates: number;
  totalApplications: number;
  byStage: StageCount[];
}